fix: lift remaining business logic from the libsolv wrapper
