+++ crates/rattler_solve/src/libsolv/wrapper/transaction.rs
@@ -1,12 +1,7 @@
 use super::ffi;
-use super::pool::{Pool, StringId};
-use super::repo::RepoId;
-use super::solvable::{self, SolvableId};
+use super::solvable::SolvableId;
 use super::solver::Solver;
 use crate::libsolv::wrapper::queue::QueueRef;
-use crate::package_operation::{PackageOperation, PackageOperationKind};
-use rattler_conda_types::RepoDataRecord;
-use std::collections::HashMap;
 use std::marker::PhantomData;
 use std::ptr::NonNull;
 
@@ -44,12 +39,12 @@ impl Transaction<'_> {
     }
 
     /// Returns a reference to the wrapped `ffi::Transaction`.
-    fn as_ref(&self) -> &ffi::Transaction {
+    pub fn as_ref(&self) -> &ffi::Transaction {
         unsafe { self.0.as_ref() }
     }
 
     /// Returns the transaction type
-    fn transaction_type(&self, solvable_id: SolvableId) -> ffi::Id {
+    pub fn transaction_type(&self, solvable_id: SolvableId) -> ffi::Id {
         unsafe {
             ffi::transaction_type(
                 self.raw_ptr(),
@@ -59,112 +54,17 @@ impl Transaction<'_> {
         }
     }
 
-    /// Returns the package operations derived from the transaction
+    /// Returns the second solvable associated to the transaction
     ///
-    /// If the transaction contains libsolv operations that have no mapping to `PackageOperation`,
-    /// an error is returned containing their ids
-    pub fn get_package_operations(
-        &mut self,
-        pool: &Pool,
-        repo_mapping: &HashMap<RepoId, usize>,
-        repodata_records: &[&[RepoDataRecord]],
-    ) -> Result<Vec<PackageOperation>, Vec<ffi::Id>> {
-        let mut package_operations = Vec::new();
-        let mut unsupported_operations = Vec::new();
-
-        let solvable_index_id = pool
-            .find_interned_str("solvable:repodata_record_index")
-            .unwrap();
-
-        let transaction = self.as_ref();
-
-        // Safe because `transaction.steps` is an active queue
-        let transaction_queue = unsafe { QueueRef::from_ffi_queue(transaction, transaction.steps) };
-
-        for id in transaction_queue.iter() {
-            let id = SolvableId(id);
-            let transaction_type = self.transaction_type(id);
-
-            // Retrieve the repodata record corresponding to this solvable
-            let (repo_index, solvable_index) =
-                get_solvable_indexes(pool, repo_mapping, solvable_index_id, id);
-            let repodata_record = &repodata_records[repo_index][solvable_index];
-
-            match transaction_type as u32 {
-                ffi::SOLVER_TRANSACTION_DOWNGRADED
-                | ffi::SOLVER_TRANSACTION_UPGRADED
-                | ffi::SOLVER_TRANSACTION_CHANGED => {
-                    package_operations.push(PackageOperation {
-                        package: repodata_record.clone(),
-                        kind: PackageOperationKind::Remove,
-                    });
-
-                    // Safe because the provided id is valid
-                    let solvable_offset =
-                        unsafe { ffi::transaction_obs_pkg(self.raw_ptr(), id.into()) };
-
-                    let second_solvable_id = SolvableId(solvable_offset);
-                    let (second_repo_index, second_solvable_index) = get_solvable_indexes(
-                        pool,
-                        repo_mapping,
-                        solvable_index_id,
-                        second_solvable_id,
-                    );
-                    let second_repodata_record =
-                        &repodata_records[second_repo_index][second_solvable_index];
-
-                    package_operations.push(PackageOperation {
-                        package: second_repodata_record.clone(),
-                        kind: PackageOperationKind::Install,
-                    });
-                }
-                ffi::SOLVER_TRANSACTION_REINSTALLED => {
-                    package_operations.push(PackageOperation {
-                        package: repodata_record.clone(),
-                        kind: PackageOperationKind::Reinstall,
-                    });
-                }
-                ffi::SOLVER_TRANSACTION_INSTALL => {
-                    package_operations.push(PackageOperation {
-                        package: repodata_record.clone(),
-                        kind: PackageOperationKind::Install,
-                    });
-                }
-                ffi::SOLVER_TRANSACTION_ERASE => {
-                    package_operations.push(PackageOperation {
-                        package: repodata_record.clone(),
-                        kind: PackageOperationKind::Remove,
-                    });
-                }
-                ffi::SOLVER_TRANSACTION_IGNORE => {}
-                _ => {
-                    unsupported_operations.push(transaction_type);
-                }
-            }
-        }
-
-        if !unsupported_operations.is_empty() {
-            return Err(unsupported_operations);
-        }
-
-        Ok(package_operations)
+    /// Safety: the caller must ensure the transaction has an associated solvable and that the
+    /// provided `solvable_id` is valid
+    pub unsafe fn obs_pkg(&self, solvable_id: SolvableId) -> SolvableId {
+        SolvableId(unsafe { ffi::transaction_obs_pkg(self.raw_ptr(), solvable_id.into()) })
     }
-}
 
-fn get_solvable_indexes(
-    pool: &Pool,
-    repo_mapping: &HashMap<RepoId, usize>,
-    solvable_index_id: StringId,
-    id: SolvableId,
-) -> (usize, usize) {
-    let solvable = id.resolve_raw(pool);
-    let solvable_index =
-        solvable::lookup_num(solvable.as_ptr(), solvable_index_id).unwrap() as usize;
-
-    // Safe because there are no active mutable borrows of any solvable at this stage
-    let repo_id = RepoId::from_ffi_solvable(unsafe { solvable.as_ref() });
-
-    let repo_index = repo_mapping[&repo_id];
-
-    (repo_index, solvable_index)
+    /// Returns the transaction's queue, containing a solvable id for each transaction
+    pub fn get_steps(&self) -> QueueRef {
+        // Safe because the transaction is live and `transaction.steps` is a queue
+        unsafe { QueueRef::from_ffi_queue(self, self.as_ref().steps) }
+    }
 }
